fix: address second round of Copilot review comments
- README.md: fix db.system → db.system.name in database span table
- otel.go: call otel.Handle(err) on instrument creation errors instead
  of silently discarding them (doc comment already promised this)
- migrate.go: use attribute.Int64 for uint version fields to avoid
  int overflow on 32-bit platforms
- migrate.go: sanitize database.Error in otelSpanSetError to avoid
  leaking migration SQL into trace span status descriptions
- database/oteldriver.go: same SQL-leak fix in endSpan; add errors import
- source/oteldriver.go: use attribute.Int64 for uint version fields
- source/oteldriver_test.go: assert span.Status().Code != codes.Error
  directly rather than comparing full sdktrace.Status struct
- source/pkger/pkger.go: fix import grouping (stdlib before third-party)
- source/google_cloud_storage/storage.go: merge context into stdlib group

+++ source/oteldriver.go
@@ -75,7 +75,7 @@ func (d *OTelDriver) Next(ctx context.Context, version uint) (uint, error) {
 func (d *OTelDriver) ReadUp(ctx context.Context, version uint) (io.ReadCloser, string, error) {
 	ctx, span := d.tracer.Start(ctx, "source.read_up",
 		trace.WithSpanKind(trace.SpanKindInternal),
-		trace.WithAttributes(append(d.attrs(), attribute.Int("migrate.version", int(version)))...),
+		trace.WithAttributes(append(d.attrs(), attribute.Int64("migrate.version", int64(version)))...),
 	)
 	r, identifier, err := d.driver.ReadUp(ctx, version)
 	if err != nil && !errors.Is(err, os.ErrNotExist) {
@@ -89,7 +89,7 @@ func (d *OTelDriver) ReadUp(ctx context.Context, version uint) (io.ReadCloser, s
 func (d *OTelDriver) ReadDown(ctx context.Context, version uint) (io.ReadCloser, string, error) {
 	ctx, span := d.tracer.Start(ctx, "source.read_down",
 		trace.WithSpanKind(trace.SpanKindInternal),
-		trace.WithAttributes(append(d.attrs(), attribute.Int("migrate.version", int(version)))...),
+		trace.WithAttributes(append(d.attrs(), attribute.Int64("migrate.version", int64(version)))...),
 	)
 	r, identifier, err := d.driver.ReadDown(ctx, version)
 	if err != nil && !errors.Is(err, os.ErrNotExist) {
